+++ src/algorithm/GreaterByN1BComparator.java
@@ -4,14 +4,14 @@ import java.util.Comparator;
 
 import model.Graph;
 
-public class GreaterByN1BComparator implements Comparator<Long> {
+public class GreaterByN1BComparator implements Comparator<Integer> {
 	private Graph g;
 
 	@Override
-	public int compare(Long arg0, Long arg1) {
+	public int compare(Integer arg0, Integer arg1) {
 		int is = g.getN1(arg0).size();
 		int js = g.getN1(arg1).size();
-		return  ((js == is) ? (int)(arg1 - arg0) : js - is);
+		return  ((js == is) ? (arg1 - arg0) : js - is);
 	}
 
 	public GreaterByN1BComparator(Graph g) {
