+++ src/algorithm/LessByN1BComparator.java
@@ -4,14 +4,14 @@ import java.util.Comparator;
 
 import model.Graph;
 
-public class LessByN1BComparator implements Comparator<Long> {
+public class LessByN1BComparator implements Comparator<Integer> {
 	private Graph g;
 
 	@Override
-	public int compare(Long arg0, Long arg1) {
+	public int compare(Integer arg0, Integer arg1) {
 		int is = g.getN1(arg0).size();
 		int js = g.getN1(arg1).size();
-		return  ((is == js) ? (int)(arg0 - arg1) : is - js);
+		return  ((is == js) ? (arg0 - arg1) : is - js);
 	}
 
 	public LessByN1BComparator(Graph g) {
