+++ src/algorithm/chapter7/Algorithm34State.java
@@ -6,43 +6,43 @@ import java.util.LinkedHashSet;
 import model.Graph;
 
 public class Algorithm34State {
-	public final Long v;
-	public LinkedHashSet<Long> W;
-	public Long w = 0L;
-	public LinkedHashSet<Long> dist2NotSorG;
-	public LinkedHashMap<Long, Long> spans;
+	public final Integer v;
+	public LinkedHashSet<Integer> W;
+	public Integer w = 0;
+	public LinkedHashSet<Integer> dist2NotSorG;
+	public LinkedHashMap<Integer, Integer> spans;
 	public final Object canJoinLock = new Object();
-	public Long canJoin = 0L;
+	public Integer canJoin = 0;
 
-	public Algorithm34State(Long forV, Graph g) {
+	public Algorithm34State(Integer forV, Graph g) {
 		this.v = forV;
 		this.W = new LinkedHashSet<>(g.getN1(forV));
 		this.dist2NotSorG = new LinkedHashSet<>(g.getN2(forV));
 		this.spans = new LinkedHashMap<>();
 	}
 
-	public void recieveSpan(Long from, Long value) {
+	public void recieveSpan(Integer from, Integer value) {
 		synchronized (spans) {
 			spans.put(from, value);
 		}
 		return;
 	}
 
-	public void recieveRemoveFromW(Long v) {
+	public void recieveRemoveFromW(Integer v) {
 		synchronized (W) {
 			W.remove(v);
 		}
 		return;
 	}
 
-	public void recieveRemoveFromSpans(Long v) {
+	public void recieveRemoveFromSpans(Integer v) {
 		synchronized (spans) {
 			spans.remove(v);
 		}
 		return;
 	}
 
-	public void recieveRemoveFromDist2(Long v) {
+	public void recieveRemoveFromDist2(Integer v) {
 		synchronized (dist2NotSorG) {
 			dist2NotSorG.remove(v);
 		}
